Ændret eventPage til at logikken ligger i FriendListModal
